Move logseq code into separate files
